Changed URLString parameter pass occasions in functions
In those functions which have url in the parameters, the type of the url was changed to URLString, but in some cases the parameter pass was by value and this parameter was only read in the function. These cases were fixed: if the value was only read, then the parameter pass was changed to const reference handover.